+++ lib/src/browser_web_auth.dart
@@ -1,6 +1,4 @@
-// ignore_for_file: avoid_web_libraries_in_flutter
-
-import 'dart:html' as html;
+import 'package:web/web.dart' as web;
 
 import 'base_web_auth.dart';
 
@@ -14,16 +12,14 @@ class BrowserWebAuth implements BaseWebAuth {
       required String redirectUrl,
       Map<String, dynamic>? opts}) async {
     // ignore: unsafe_html
-    final popupLogin = html.window.open(
-        url,
-        'oauth2_client::authenticateWindow',
+    final popupLogin = web.window.open(url, 'oauth2_client::authenticateWindow',
         'menubar=no, status=no, scrollbars=no, menubar=no, width=1000, height=500');
 
-    var messageEvt = await html.window.onMessage
+    var messageEvt = await web.window.onMessage
         .firstWhere((evt) => evt.origin == Uri.parse(redirectUrl).origin);
 
-    popupLogin.close();
+    popupLogin?.close();
 
-    return messageEvt.data;
+    return messageEvt.data.toString();
   }
 }